wrapper: Move self-contained functions into utils.py
