+++ wrapper/common.py
@@ -4,7 +4,6 @@ import logging
 import os
 import re
 import signal
-import stat
 import subprocess
 import sys
 import tempfile
@@ -123,24 +122,6 @@ def write_password(password, host):
     os.write(pfile[0], bytes(password.encode('utf-8')))
     os.close(pfile[0])
     return pfile[1]
-
-
-def add_perms_to_file(path, modes, uid=-1, gid=-1):
-    cur_mode = stat.S_IMODE(os.stat(path).st_mode)
-    new_mode = cur_mode | modes
-
-    if uid != -1 or gid != -1:
-        logging.debug('Changing uid:gid of "%s" to %s:%s',
-                      path, uid, gid)
-        os.chown(path, uid, gid)
-
-    logging.debug('Changing permissions on "%s" from 0%o to 0%o',
-                  path, cur_mode, new_mode)
-    os.chmod(path, new_mode)
-
-
-def nbd_uri_from_unix_socket(sock_path):
-    return 'nbd+unix:///?socket=%s' % sock_path
 
 
 def setup_signals():
